+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Support for `logo` field in `EmailTemplate` class to specify a custom logo URL for booking emails
 * Support for `show_nylas_branding` field in `EmailTemplate` class to control Nylas branding visibility in booking emails
 * Support for `metadata` field type in `AdditionalFieldType` enum for scheduler additional fields
+* Support for `include_hidden_folders` query parameter in `ListFoldersQueryParams` for Microsoft accounts to control whether hidden folders are included in the response
 
 ## [2.10.0] - Release 2025-06-12